feat(sidebar): replace NavProject with NavResources in AppSidebar for better resource management feat(nav): add NavResources component to manage project resources refactor(ui): create UiBanner and UiContainer components for consistent layout refactor(ui): create UiSideBarItem for sidebar item rendering refactor(ui): create UiTable for table rendering with react-table integration chore: remove deprecated WorkspacesTable component and related code feat(routes): update project routes to use UiBanner and UiContainer for better UI consistency
