+++ client/src/routes/projects/media/index.tsx
@@ -1,12 +1,14 @@
-import {Outlet} from 'react-router';
+import {UiBanner} from '@/components/ui/UiBanner';
+import {UiContainer} from '@/components/ui/UiContainer';
 import type {Route} from './+types/index';
 
 export default function Index({params}: Route.LoaderArgs) {
     return (
-        <div>
-            <h1>Media Index</h1>
-            <div>This page lists all media under a project</div>
-            <Outlet />
-        </div>
+        <UiContainer>
+            <UiBanner
+                title="Medias"
+                description="Here's a list of your medias for this project!"
+            />
+        </UiContainer>
     );
 }
