+++ client/src/routes/projects/workspaces/index.tsx
@@ -1,22 +1,37 @@
-import {WorkspacesTable} from '@/components/workspaces/WorkspacesTable';
-import {Outlet} from 'react-router';
+import type {WorkspaceDto} from '@/api';
+import {UiBanner} from '@/components/ui/UiBanner';
+import {UiContainer} from '@/components/ui/UiContainer';
+import {UiTable} from '@/components/ui/UiTable';
+import type {ColumnDef} from '@tanstack/react-table';
 import type {Route} from './+types/index';
+
+export const COLUMNS: ColumnDef<WorkspaceDto>[] = [
+    {
+        accessorKey: 'id',
+        header: 'ID'
+    },
+    {
+        accessorKey: 'engine',
+        header: 'Engine'
+    },
+    {
+        accessorKey: 'name',
+        header: 'Name'
+    },
+    {
+        accessorKey: 'description',
+        header: 'Description'
+    }
+];
 
 export default function Index({params}: Route.LoaderArgs) {
     return (
-        <div className="flex flex-col mx-auto w-full max-w-6xl py-10 gap-5">
-            <div className="flex items-center justify-between space-y-2">
-                <div>
-                    <h2 className="text-2xl font-bold tracking-tight">
-                        Workspaces
-                    </h2>
-                    <p className="text-muted-foreground">
-                        Here&apos;s a list of your workspaces for this project!
-                    </p>
-                </div>
-            </div>
-            <WorkspacesTable />
-            <Outlet />
-        </div>
+        <UiContainer>
+            <UiBanner
+                title="Workspaces"
+                description="Here's a list of your workspaces for this project!"
+            />
+            <UiTable columns={COLUMNS} data={[]} />
+        </UiContainer>
     );
 }
